TransferFunctionEditor, HasTransferFunction: move tf and display range IO to HasTransferFunction interface.
